requirements_exe.txt
PyInstaller>=5.0
pystray>=0.19.4
Pillow>=9.0.0
win10toast>=0.9
edge-tts>=6.1.0
pygame>=2.1.0
pyttsx3>=2.90
requests>=2.28.0
flask>=2.2.0
flask-socketio>=5.3.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
pynput>=1.7.6
numpy>=1.21.0
python-dotenv>=1.0.0 